Move quota to daemon/internal/quota
