+++ daemon/graphdriver/graphtest/graphtest_unix.go
@@ -11,8 +11,8 @@ import (
 	"testing"
 
 	"github.com/docker/docker/daemon/graphdriver"
+	"github.com/docker/docker/daemon/internal/quota"
 	"github.com/docker/docker/pkg/stringid"
-	"github.com/docker/docker/quota"
 	"github.com/docker/go-units"
 	"golang.org/x/sys/unix"
 	"gotest.tools/v3/assert"
