+++ daemon/graphdriver/overlay2/overlay.go
@@ -20,9 +20,9 @@ import (
 	"github.com/docker/docker/daemon/graphdriver/overlayutils"
 	"github.com/docker/docker/daemon/internal/fstype"
 	"github.com/docker/docker/daemon/internal/mountref"
+	"github.com/docker/docker/daemon/internal/quota"
 	"github.com/docker/docker/internal/containerfs"
 	"github.com/docker/docker/internal/directory"
-	"github.com/docker/docker/quota"
 	"github.com/docker/go-units"
 	"github.com/moby/go-archive"
 	"github.com/moby/go-archive/chrootarchive"
