+++ daemon/volume/local/local_unix.go
@@ -14,8 +14,8 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/docker/docker/daemon/internal/quota"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/quota"
 	"github.com/docker/go-units"
 	"github.com/moby/sys/mount"
 	"github.com/moby/sys/mountinfo"
